streamer/requirements.txt
pyyaml
numpy
#vimba
pymba
opencv-python
pyserial
pynmea2
